requirements.txt
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
opencv-python>=4.5.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0


